Add lifecycle to demo project and Delete separate Lifecycle project
