+++ Demo/Demo/ViewController.swift
@@ -17,6 +17,8 @@ class ViewController: UIViewController {
     
     override func viewDidLoad() {
         super.viewDidLoad()
+        print("\(#function) State: \(UIApplication.shared.applicationState.toString())")
+        // Do any additional setup after loading the view.
         
         let initialVal = Bundle.main.infoDictionary?["InitialProgress"] as? Int
         updateCounter(initialVal ?? 50)
@@ -24,6 +26,30 @@ class ViewController: UIViewController {
         let label = UILabel()
         view.addSubview(label)
         // Do any additional setup after loading the view.
+    }
+    
+    override func loadView() {
+        super.loadView()
+        print("\(#function) State: \(UIApplication.shared.applicationState.toString())")
+    }
+    
+    override func viewWillAppear(_ animated: Bool) {
+        super.viewWillAppear(animated)
+        print("\(#function) State: \(UIApplication.shared.applicationState.toString())")
+    }
+    
+    override func viewDidAppear(_ animated: Bool) {
+        super.viewDidAppear(animated)
+        print("\(#function) State: \(UIApplication.shared.applicationState.toString())")
+    }
+    
+    override func viewWillDisappear(_ animated: Bool) {
+        super.viewWillDisappear(animated)
+        print("\(#function) State: \(UIApplication.shared.applicationState.toString())")
+    }
+    
+    deinit {
+        print("\(#function) State: \(UIApplication.shared.applicationState.toString())")
     }
     
     @IBAction func reset(_ sender: UIButton) {
